+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Help text that shows with `--help`.
+
 ## [0.1.3] - 2020-12-16
 
 ### Added