+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2020-12-23
+
 ### Added
 
 - Help text that shows with `--help`.
 - Fix whole directories.
+- Fix directories and subdirectories recursivly.
+
+### Changed
+
+- Major refactorings of production and test code.
 
 ## [0.1.3] - 2020-12-16
 
@@ -45,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/thled/namespacer/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/thled/namespacer/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/thled/namespacer/compare/v0.1.3...v1.0.0
 [0.1.3]: https://github.com/thled/namespacer/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/thled/namespacer/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/thled/namespacer/compare/v0.1.0...v0.1.1